+++ RELEASE_NOTES.md
@@ -16,7 +16,7 @@
 
 #### 0.0.2 - November 14, 2017
 * standardize on verb "Create" instead of "Parse"
-* Create overrides for LimitedValue
+* Create overrides for DependentType
 
 #### 0.0.1 - October 28, 2017
 * Initial release